Client: renamed quantum_function file to program
