+++ client/quantum_serverless/core/provider.py
@@ -50,7 +50,7 @@ from quantum_serverless.core.job import (
     GatewayJobClient,
     BaseJobClient,
 )
-from quantum_serverless.core.quantum_function import QuantumFunction
+from quantum_serverless.core.program import QuantumFunction
 from quantum_serverless.core.tracing import _trace_env_vars
 from quantum_serverless.exception import QuantumServerlessException
 from quantum_serverless.utils import JsonSerializable
