+++ client/tests/core/test_quantum_function_repository.py
@@ -5,7 +5,7 @@ import shutil
 from pathlib import Path
 from unittest import TestCase, mock
 
-from quantum_serverless.core.quantum_function import (
+from quantum_serverless.core.program import (
     QuantumFunctionRepository,
     QuantumFunction,
 )
